pyproject.toml
[project]
name = "category"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "decouple>=0.0.7",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-stubs>=5.2.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-stubs>=3.16.0",
    "drf-nested-routers>=0.94.2",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.4.0",
]
